requirements.txt
streamlit>=1.40.0
google-genai>=1.2.0
requests
confluent-kafka
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
